Set -np in mymac xml file to 1
There's no benefit to setting --np to a value greater
than 1 in single point mode for the OLMT. (In fact,
model will crash if trying to distribute ntasks in
env_mach_pes.xml beyond 1).

If running a larger domain, might want to modify this file
as well as env_mach_pes.xml of your case.